worker/requirements.txt
flask
google-genai
gspread
google-auth
python-dotenv
yt-dlp
httpx
edge-tts
Pillow
mutagen
